Seperated eventapi from api.py
